+++ .size-limit.js
@@ -7,19 +7,19 @@ module.exports = [
     name: 'Core - CDN',
     path: 'dist/legacy/rudder-analytics.min.js',
     gzip: true,
-    limit: '41.3 kB',
+    limit: '40.9 kB',
   },
   {
     name: 'All Integrations - CDN',
     path: 'dist/legacy/js-integrations/*.min.js',
     gzip: true,
-    limit: '463.6 kB',
+    limit: '444.7 kB',
   },
   {
     name: 'Core - NPM',
     path: 'dist/npm-lib/index.js',
     gzip: true,
-    limit: '41.3 kB',
+    limit: '40.8 kB',
   },
   {
     name: 'Service Worker - NPM',