+++ src/integrations/Mixpanel/browser.js
@@ -21,7 +21,6 @@
 /* eslint-disable vars-on-top */
 /* eslint-disable one-var */
 /* eslint-disable no-underscore-dangle */
-/* eslint-disable prettier/prettier */
 /* eslint-disable class-methods-use-this */
 import get from 'get-value';
 import logger from '../../utils/logUtil';
@@ -77,9 +76,12 @@ class Mixpanel {
       username: '$username',
       phone: '$phone',
     };
-    this.areTransformationsConnected =
-      destinationInfo && destinationInfo.areTransformationsConnected;
-    this.destinationId = destinationInfo && destinationInfo.destinationId;
+    this.identityMergeApi = config.identityMergeApi || 'original';
+    ({
+      shouldApplyDeviceModeTransformation: this.shouldApplyDeviceModeTransformation,
+      propagateEventsUntransformedOnError: this.propagateEventsUntransformedOnError,
+      destinationId: this.destinationId,
+    } = destinationInfo ?? {});
   }
 
   init() {
@@ -187,12 +189,15 @@ class Mixpanel {
     peopleProperties = extendTraits(peopleProperties);
     const superProperties = parseConfigArray(this.superProperties, 'property');
 
-    // eslint-disable-next-line camelcase
-    const user_id = rudderElement.message.userId || rudderElement.message.anonymousId;
+    let userId = rudderElement.message.userId || rudderElement.message.anonymousId;
+    if (this.identityMergeApi === 'simplified') {
+      // calling mixpanel .identify() only for known users
+      userId = rudderElement.message.userId;
+    }
     let traits = formatTraits(rudderElement.message);
     const { email, username } = traits;
     // id
-    if (user_id) window.mixpanel.identify(user_id);
+    if (userId) window.mixpanel.identify(userId);
 
     // name tag
     const nametag = email || username;
@@ -363,7 +368,7 @@ class Mixpanel {
      * groupIdentifierTraits: [ {trait: "<trait_value>"}, ... ]
      */
     const identifierTraitsList = parseConfigArray(this.groupKeySettings, 'groupKey');
-    if (traits && Object.keys(traits).length) {
+    if (traits && Object.keys(traits).length > 0) {
       identifierTraitsList.forEach((trait) => {
         window.mixpanel.get_group(trait, groupId).set_once(traits);
       });
@@ -372,25 +377,32 @@ class Mixpanel {
   }
 
   /**
+   * https://github.com/mixpanel/mixpanel-js/blob/master/doc/readme.io/javascript-full-api-reference.md#mixpanelalias
    * @param {*} rudderElement
    */
   alias(rudderElement) {
     logger.debug('in Mixpanel alias');
+    if (this.identityMergeApi === 'simplified') {
+      logger.debug("===Mixpanel: Alias call is deprecated in 'Simplified ID Merge'===");
+      return;
+    }
+
     const { previousId, userId } = rudderElement.message;
+    const newId = userId;
     if (!previousId) {
       logger.debug('===Mixpanel: previousId is required for alias call===');
       return;
     }
-    if (!userId) {
+    if (!newId) {
       logger.debug('===Mixpanel: userId is required for alias call===');
       return;
     }
 
-    if (window.mixpanel.get_distinct_id && window.mixpanel.get_distinct_id() === userId) {
+    if (window.mixpanel.get_distinct_id && window.mixpanel.get_distinct_id() === newId) {
       logger.debug('===Mixpanel: userId is same as previousId. Skipping alias ===');
       return;
     }
-    window.mixpanel.alias(userId, previousId);
+    window.mixpanel.alias(newId, previousId);
   }
 }
 export default Mixpanel;